added SearchBar and RandomButton components for better readability from the header
