+++ src/components/Header/Header.css
@@ -1,7 +1,7 @@
 .header {
   display: flex;
   margin: 0;
-  
+
   width: 100%;
   background-color: #416165;
 }
@@ -17,136 +17,17 @@
   z-index: 1000;
 }
 
-.header__btn-container {
-  position: relative;
-  bottom: 2px;
-}
 .header_centered {
   display: flex;
   flex-direction: column;
   align-items: center;
   justify-content: center;
-  height: 100vh; /* Full viewport height */
+  height: 100vh;
   text-align: center;
 }
 
-.header__searchbar_centered {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 15px;
-  width: 100%;
-  max-width: 500px; /* Similar width to Google's search bar */
-}
-
-.header__searchbar_top {
-  display: flex;
-  height: 36px;
-  gap: 10px;
-  
-
-}
-
-.header__search_centered {
-  width: 100%;
-  padding: 12px 16px;
-  font-size: 16px;
-  border: 1px solid #dfe1e5;
-  border-radius: 24px;
-  outline: none;
-  transition: box-shadow 0.2s ease-in-out;
-
-}
-.header__search_top {
-  position: relative;
-  top: 2px;
-  border: 1px solid #dfe1e5;
-  border-radius: 24px;
-  outline: none;
-  transition: box-shadow 0.2s ease-in-out;
-  padding: 12px 16px;
-}
-
-.header__search:focus {
-  box-shadow: 0px 0px 5px rgba(0, 0, 0, 0.2);
-}
-
-.header__dropdown-container {
-  display: inline-block;
-  
-}
-
-.header__randomAnimeBtn {
-  height: 36px;
-  padding: 8px 16px;
-  border: none;
-  border-radius: 4px 0 0 4px; /* Rounded only on the left */
-  background-color: #4285f4;
-  color: white;
-  cursor: pointer;
-  transition: background-color 0.3s;
-}
-
-.header__dropdown-btn {
-  position: relative;
-  height: 36px;
-  top: 3px;
-  border-radius: 0 4px 4px 0;
-  border-color: transparent;
-  cursor: pointer;
-  margin-right: 40px;
-}
-
-.header__randomAnimeBtn:hover {
-  background-color: #357ae8;
-}
-
-.header__dropdown {
-  position: absolute;
-  background-color: white;
-  border: 1px solid #ccc;
-  border-radius: 5px;
-  max-height: 200px;
-  width: 150px;
-  overflow-y: auto;
-  padding: 5px;
-  box-shadow: 0px 2px 5px rgba(0, 0, 0, 0.2);
-  z-index: 999;
-  cursor: pointer;
-  right: 40px;
-  
-}
-
-.dropdown-option {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  padding: 5px;
-}
-
-.dropdown-option input {
-  cursor: pointer;
-}
-
-.dropdown-option label {
-  cursor: pointer;
-  font-size: 14px;
-}
-
-.dropdown_img {
-  position: relative;
-  height: 20px;
-}
 .header__title {
   cursor: pointer;
   color: #d9dbf1;
   margin: 0;
 }
-
-/* .header__randomAnimeBtn {
-
-}
-
-.header__dropdown-btn {
-
-} */
